Upstream version 2.73+dfsg0

# gpg: Signature made Thu Jan  8 14:31:28 2015 CET
# gpg:                using RSA key 0x0612F4944ECDCD5A
# gpg: Good signature from "Matteo F. Vescovi <mfv@debian.org>" [ultimate]
# gpg:                 aka "Matteo F. Vescovi <mfvescovi@gmail.com>" [ultimate]
# Primary key fingerprint: 4E8E 810A 6B44 5FDE 68DA  D025 8062 3989 83B2 CF7A
#      Subkey fingerprint: F3DD 09F8 DA87 D1DF 5046  C99B 0612 F494 4ECD CD5A

+++ source/blender/editors/sculpt_paint/paint_image_2d.c
@@ -38,9 +38,7 @@
 #include "DNA_space_types.h"
 #include "DNA_object_types.h"
 
-#include "BLI_math.h"
 
-#include "BLI_rect.h"
 #include "BLI_math_color_blend.h"
 #include "BLI_stack.h"
 #include "BLI_bitmap.h"
@@ -65,7 +63,6 @@
 
 #include "UI_view2d.h"
 
-#include "RE_shader_ext.h"
 
 #include "GPU_draw.h"
 
@@ -1564,7 +1561,7 @@ void paint_2d_gradient_fill(
 	sub_v2_v2v2(tangent, image_final, image_init);
 	line_len = len_squared_v2(tangent);
 	line_len_sq_inv = 1.0f / line_len;
-	line_len = sqrt(line_len);
+	line_len = sqrtf(line_len);
 
 	do_float = (ibuf->rect_float != NULL);
